Refactor employers and participants services
