+++ server/services/employers.js
@@ -1,11 +1,8 @@
-const { getUserRoles, getUserRegionsCriteria } = require('./user.js');
 const { dbClient, collections } = require('../db');
+const { userRegionQuery } = require('./user.js');
 
-const getEmployers = async (req) => {
-  const roles = getUserRoles(req);
-  const isMOH = roles.includes('ministry_of_health');
-  const isSuperUser = roles.includes('superuser');
-  const criteria = isSuperUser || isMOH ? {} : getUserRegionsCriteria(req, 'healthAuthority');
+const getEmployers = async (user) => {
+  const criteria = user.isSuperUser || user.isMOH ? {} : userRegionQuery(user.regions, 'healthAuthority');
   return criteria
     ? dbClient.db[collections.EMPLOYER_FORMS].findDoc(criteria)
     : [];
